--- QBert_Project/Assets/Assets/Animation/Qbert Animations.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: f6b719a96692e4422ad92afa353f2f58
-folderAsset: yes
-timeCreated: 1522686663
-licenseType: Pro
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
