--- QBert_Project/Assets/CoilyAnimationScript.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CoilyAnimationScript : MonoBehaviour {
-
-    public void DestroyAnimation(){
-        CoilyScript.Move();
-        Destroy(gameObject);
-    }
-}
