+++ QBert_Project/Assets/Scripts/CoilyAnimationScript.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class CoilyAnimationScript : MonoBehaviour {
+
+    public void DestroyAnimation(){
+        CoilyScript.Move();
+        Destroy(gameObject);
+    }
+}
