+++ QBert_Project/ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2017.3.0f3
+m_EditorVersion: 2017.3.1f1
